+++ DiskLayer/DiskManager.h
@@ -11,7 +11,7 @@
 #include <unistd.h>
 #include <fstream>
 #include <iostream>
-#include "Page.h"
+#include "../Models/Page.h"
 
 inline std::string filepath = "/Users/markseeliger/CLionProjects/marksql/Storage/temp.db";
 //0644 is the file permissions that grants the owner read and write, and everyone else read only
@@ -22,16 +22,16 @@ class DiskManager {
 private:
     std::string filepath_;
     int fd_;
-    size_t next_page_id_;
+    int next_page_id_;
 public:
     DiskManager();
     explicit DiskManager(std::string customFilepath);
     char* readPage(int page_id, char* buf);
     void writePage(int page_id, char* buf);
 
-    size_t allocatePage();
+    int allocatePage();
 
-    static off_t getByteOffset(size_t page_id);
+    static int getByteOffset(int page_id);
     ~DiskManager();
 
 private:
